requirements.txt
# NL2PD Backend Requirements
# Complete dependencies for the FastAPI backend application

# Core FastAPI and ASGI server
fastapi==0.116.1
uvicorn[standard]==0.35.0
starlette==0.47.2

# Data manipulation and analysis
pandas==2.3.1
numpy==2.3.2

# File handling
openpyxl==3.1.5
python-multipart==0.0.20

# HTTP requests and API communication
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
h2==4.2.0

# Authentication and security
PyJWT==2.10.1
cryptography==45.0.6

# Environment variable management
python-dotenv==1.1.1

# Pydantic for data validation
pydantic==2.11.7
pydantic-core==2.33.2

# Type checking and annotations
typing-extensions==4.14.1

# SSL/TLS certificate handling
certifi==2025.8.3

# Database integration (Supabase)
supabase==2.18.1
postgrest==1.1.1
supabase-auth==2.12.3
supabase-functions==0.10.1
storage3==0.12.1
realtime==2.7.0

# AI integration
groq==0.31.0
distro==1.9.0

# Fuzzy string matching for entity classification
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
Levenshtein==0.27.1
rapidfuzz==3.13.0

# Additional utilities and dependencies
python-dateutil==2.9.0.post0
click==8.2.1
colorama==0.4.6
idna==3.10
charset-normalizer==3.4.3
urllib3==2.5.0
websockets==15.0.1
packaging==25.0
pytz==2025.2
tzdata==2025.2
six==1.17.0
sniffio==1.3.1
annotated-types==0.7.0
anyio==4.10.0
cffi==1.17.1
pycparser==2.22
deprecation==2.1.0
et-xmlfile==2.0.0
hpack==4.1.0
hyperframe==6.1.0
StrEnum==0.4.15
typing-inspection==0.4.1
pytest-asyncio==0.25.0

# Development tools
httpx==0.28.1
